fix world totals
