+++ scripts/analysis/common.py
@@ -52,12 +52,15 @@ def create_world_total(data: pd.DataFrame, name: str = "World") -> pd.DataFrame:
 
     df = data.copy(deep=True)
     df["country"] = name
-    df = df.groupby(
-        [c for c in df.columns if c not in ["income_level", "continent"]],
-        observed=True,
-        dropna=False,
-        as_index=False,
-    )["value"].sum()
+    df = (
+        df.groupby(
+            [c for c in df.columns if c not in ["income_level", "continent", "value"]],
+            observed=True,
+            dropna=False,
+        )["value"]
+        .sum()
+        .reset_index()
+    )
 
     return pd.concat([data, df], ignore_index=True)
 
@@ -76,3 +79,67 @@ def add_china_as_counterpart_type(df: pd.DataFrame) -> pd.DataFrame:
 
     # Concatenate the data
     return pd.concat([df, china], ignore_index=True)
+
+
+def convert_to_net_flows(data: pd.DataFrame) -> pd.DataFrame:
+    """Group the indicator type to get net flows"""
+
+    data = (
+        data.groupby(
+            [c for c in data.columns if c not in ["value", "indicator_type"]],
+            observed=True,
+            dropna=False,
+        )["value"]
+        .sum()
+        .reset_index()
+    )
+
+    data["indicator_type"] = "net_flow"
+
+    return data
+
+
+def summarise_by_country(data: pd.DataFrame) -> pd.DataFrame:
+    """Summarise the data by country"""
+
+    data = (
+        data.groupby(
+            [
+                c
+                for c in data.columns
+                if c
+                not in [
+                    "value",
+                    "counterpart_area",
+                    "counterpart_type",
+                    "indicator",
+                ]
+            ],
+            observed=True,
+            dropna=False,
+        )["value"]
+        .sum()
+        .reset_index()
+    )
+
+    return data
+
+
+def create_groupings(data: pd.DataFrame) -> pd.DataFrame:
+    # Create world totals
+    data_grouped = create_world_total(data, "Developing countries")
+
+    # Create continent totals
+    data_grouped = create_grouping_totals(
+        data_grouped, group_column="continent", exclude_cols=["income_level"]
+    )
+
+    # Create income_level totals
+    data_grouped = create_grouping_totals(
+        data_grouped, group_column="income_level", exclude_cols=["continent"]
+    )
+
+    # remove individual country data
+    data_grouped = data_grouped.loc[lambda d: d.country.isin(GROUPS)]
+
+    return data_grouped
